slow progress in protocode
Ok so Ι cant use temporal-functions as they are right now as, with input
events, we only care about time according to the events themselves.

Instead Im gonna make a macro very much of the same flavor as
temporal-functions but much simpler.

Ther other thing is that we need both, the full input state AND the
input itself at all times for combos. To this end we alloww you to
listen to the slots of the kind, so you listen to ALL mouse button
changes for example (but would recieve none of the mouse move events)

Every event-source now has a 'containers slot. This holds the array the
source is currently inside with the idea that when the source
is updated (via #'apply-*) we can propagate easily.

This is looking ok but now want to minimize combos and work on their
propagate.