Finally got a nice propagate api
You now get the data first and then the input source so you always query
what you need (or just use the data as is if thats all you wanted).

We don't expose the control object to the user at all, and require
everything to go through the input source.

We have moved defcombo into protocode as I need to make some actually
projects with button combos in order to validate the approach.