Worked out how I want the predicate-sources to be
Renamed 'em to logicial-control, make it like a cond form where you can
specialize on events from different kinds of control.

A logicial-control is now much more like a regular control, in that it
must be attached to a input-source for it to work.

event-listeners are now a more primitive type, they are not a kind of
logicial-control (whereas they used to be a kind of predicate-source)

controls can now be static, which turns them into structs internally.

controls no longer have a common super-type (as this would mean we cant
have them be classes and structs)

I've cleaned up a bit of the symbol hiding, but there is still stuff to
do.

Input source can have one of each type of logical-control attached. This
will be changed next. It may be cool to be able to listen to an entire
array of controls.

I need to revisit the logical-control's handling of 'this' in regard to
how listeners pass this information

Finally as users never use the control objects directly we dont need
sdl2/glop to use skitter.internals as much