Add decay for controls
This is to allow controls to return to their initial value when there
is no newer input. Relative mouse motion was the first case that
needed this.

The requirement for this feature is that you call
`(next-skitter-frame)` before pushing events. This sets the
frame-counter that is used internally to determine whether to reset
the value.

If next-skitter-frame is not called the library still works but decay
will not.